requirements.txt
llama-index-embeddings-huggingface
llama-index-llms-ollama